+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.13.2] - 2025-01-20
+### Fixed
+- Issue with model_construct_recursive with optional values
+- Issue with FormLayout in List and Model fields in discriminated unions
+
+## [0.13.2] - 2025-01-20
+### Fixed
+- Working list of union in list of union
 
 ## [0.13.1] - 2025-01-19
 ### Added